DidgetGen EXE file
This is The exe File of My Project. If you just want the exe File Download This file.

Download Instructions:

-On PC On the right hand side of your screen there will be a download icon.(if you hover over it, it will say download raw file.)

-Click on that Icon and Select where you want it to Downlaod. 

-After the Download is Complete Install it.

WARNING:

Your PC will potentially Flag it as a Threat. I can assure It is 100% safe to use and edit. 
The reason it is Being Identified as a false positive in Anti Virus Software is because the EXE file And the project was made Using A VCL Framework.